Fast assets

+++ docker-entrypoint.sh
@@ -1,14 +1,17 @@
 #!/bin/bash
 set -e
 
+if [ "$1" = 'update' ]; then
+  bundle exec rake db:create db:migrate assets:precompile
+  exit 0
+fi
+
 if [ "$1" = 'worker' ]; then
   rm -rf tmp/pids/*
   bundle exec sidekiq -C config/sidekiq.yml
 fi
 
 if [ "$1" = 'web' ]; then
-  bundle exec rake db:create db:migrate
-  bundle exec rake assets:precompile
   rm -rf tmp/pids/*
   bundle exec rails server puma -p 80 --binding 0.0.0.0
 fi
@@ -22,6 +25,7 @@ if [ "$1" = 'test' ]; then
   export RAILS_ENV=test
   bundle exec rake db:create db:migrate
   bundle exec rspec
+  exit 0
 fi
 
 exec "$@"
